Add files via upload
The first video shows the indoor planning process of the BACA algorithm, the second video shows the indoor planning process of our proposed ADL-ACO algorithm, the third video shows the outdoor planning process of the BACA algorithm, and the fourth video shows the outdoor planning process of our proposed ADL-ACO algorithm. Given the file size limitation of 10MB on the upload platform, the video footage of the outdoor planning process has been subjected to a 2.5-fold acceleration. In particular, the indoor environment experiments use the Dashgo B1 mobile robot platform, which uses LiDAR as a sensor, accomplishes network communication between the host computer and a PC over a LAN, and constructs a map of the actual environment using the Gmapping algorithm, with the same parameters of ADL-ACO as in the simulation experiments. The parameters of ADL-ACO are the same as those in the simulation experiments. The outdoor environment experiments use Xavier Nx, whose algorithm parameters are the same as those of Dashgo B1. The planning process has the same start and end points and vehicle physical parameters. Since BACA does not optimize path smoothing, it suffers from lag when it encounters a turning point during operation. ADL-ACO can make the robot run smoothly to the target point.